requirements.txt
ruamel.yaml~=0.18.6
pytest~=8.3.3
pytest-mock~=3.14.0
flask~=3.0.3
flasgger~=0.9.7.1
pygithub~=2.4.0
PyJWT~=2.9.0
requests~=2.32.3
cryptography~=43.0.1
pyspellchecker==0.8.1
unidiff~=0.7.5
